envio de cambios segundo nivel
