+++ Assets/Prefabs/Enemy/Boss Skeleton.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 301f63fd3502b5146acc57deb099186d
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
